+++ src/test/java/com/iexec/core/worker/WorkerServiceTests.java
@@ -25,7 +25,10 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.*;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.MockitoAnnotations;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -33,6 +36,7 @@ import java.time.Instant;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 
+import static com.iexec.common.utils.DateTimeUtils.addMinutesToDate;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
@@ -94,7 +98,6 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .build();
 
         when(workerRepository.findByWalletAddress(walletAddress)).thenReturn(Optional.of(existingWorker));
@@ -115,7 +118,6 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .build();
 
         Worker newWorker = Worker.builder()
@@ -124,7 +126,6 @@ class WorkerServiceTests {
                 .os("otherOS")
                 .cpu("otherCpu")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .build();
 
         when(workerRepository.findByWalletAddress(walletAddress)).thenReturn(Optional.of(existingWorker));
@@ -145,7 +146,6 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .build();
         when(workerRepository.findByWalletAddress(walletAddress)).thenReturn(Optional.empty());
         when(workerRepository.save(Mockito.any())).thenReturn(worker);
@@ -154,7 +154,7 @@ class WorkerServiceTests {
         // check that the save method was called once
         Mockito.verify(workerRepository, Mockito.times(1)).save(Mockito.any());
         assertThat(addedWorker.getName()).isEqualTo(worker.getName());
-        assertThat(addedWorker.getLastAliveDate()).isEqualTo(worker.getLastAliveDate());
+        assertThat(workerService.getWorkerStatsMap().get(walletAddress).getLastAliveDate()).isBefore(new Date());
     }
 
     // isAllowedToJoin
@@ -179,7 +179,7 @@ class WorkerServiceTests {
         assertThat(workerService.isAllowedToJoin("w3")).isFalse();
     }
 
-    // updateLasAlive
+    // updateLastAlive
 
     @Test
     void shouldUpdateLastAlive() throws ParseException {
@@ -191,40 +191,17 @@ class WorkerServiceTests {
                 .id("1")
                 .name(workerName)
                 .walletAddress(walletAddress)
-                .lastAliveDate(oldLastAlive)
                 .build();
         when(workerRepository.findByWalletAddress(walletAddress)).thenReturn(Optional.of(worker));
 
         // call
-        Optional<Worker> updatedWorker = workerService.updateLastAlive(walletAddress);
-
-        // check argument passed to the save method
-        ArgumentCaptor<Worker> argument = ArgumentCaptor.forClass(Worker.class);
-        Mockito.verify(workerRepository).save(argument.capture());
-        assertThat(argument.getValue().getId()).isEqualTo(worker.getId());
-        assertThat(argument.getValue().getName()).isEqualTo(worker.getName());
-        // check that the save method was called with a lastAlive parameter updated less than a second ago
-        Date now = new Date();
-        long duration = now.getTime() - argument.getValue().getLastAliveDate().getTime();
-        long diffInSeconds = TimeUnit.MILLISECONDS.toSeconds(duration);
-        assertThat(diffInSeconds).isZero();
+        workerService.updateLastAlive(walletAddress);
 
         // check object returned by the method
-        assertThat(updatedWorker).isPresent();
-        assertThat(updatedWorker.get().getId()).isEqualTo(worker.getId());
-        assertThat(updatedWorker.get().getName()).isEqualTo(worker.getName());
-        duration = now.getTime() - updatedWorker.get().getLastAliveDate().getTime();
-        diffInSeconds = TimeUnit.MILLISECONDS.toSeconds(duration);
+        Date now = new Date();
+        long duration = now.getTime() - workerService.getWorkerStatsMap().get(walletAddress).getLastAliveDate().getTime();
+        long diffInSeconds = TimeUnit.MILLISECONDS.toSeconds(duration);
         assertThat(diffInSeconds).isZero();
-    }
-
-    @Test
-    void shouldNotFindWorkerForUpdateLastAlive() {
-        String walletAddress = "0x1a69b2eb604db8eba185df03ea4f5288dcbbd248";
-        when(workerRepository.findByWalletAddress(walletAddress)).thenReturn(Optional.empty());
-
-        Optional<Worker> optional = workerService.updateLastAlive(walletAddress);
-        assertThat(optional).isEmpty();
     }
 
     // isWorkerAllowedToAskReplicate
@@ -233,8 +210,8 @@ class WorkerServiceTests {
     void shouldWorkerBeAllowedToAskReplicate() {
         String wallet = "wallet";
         Worker worker = Worker.builder()
-                .lastReplicateDemandDate(Date.from(Instant.now().minusSeconds(10)))
                 .build();
+        workerService.getWorkerStatsMap().computeIfAbsent(wallet, WorkerService.WorkerStats::new);
         when(workerRepository.findByWalletAddress(wallet)).thenReturn(Optional.of(worker));
         when(workerConfiguration.getAskForReplicatePeriod()).thenReturn(5000L);
 
@@ -244,9 +221,7 @@ class WorkerServiceTests {
     @Test
     void shouldWorkerBeAllowedToAskReplicateSinceFirstTime() {
         String wallet = "wallet";
-        Worker worker = new Worker();
-        worker.setLastReplicateDemandDate(null);
-        when(workerRepository.findByWalletAddress(wallet)).thenReturn(Optional.of(worker));
+        workerService.getWorkerStatsMap().computeIfAbsent(wallet, WorkerService.WorkerStats::new);
 
         assertThat(workerService.isWorkerAllowedToAskReplicate(wallet)).isTrue();
     }
@@ -254,10 +229,8 @@ class WorkerServiceTests {
     @Test
     void shouldWorkerNotBeAllowedToAskReplicateSinceTooSoon() {
         String wallet = "wallet";
-        Worker worker = Worker.builder()
-                .lastReplicateDemandDate(Date.from(Instant.now().minusSeconds(1)))
-                .build();
-        when(workerRepository.findByWalletAddress(wallet)).thenReturn(Optional.of(worker));
+        workerService.getWorkerStatsMap().computeIfAbsent(wallet, WorkerService.WorkerStats::new)
+                .setLastReplicateDemandDate(Date.from(Instant.now().minusSeconds(1)));
         when(workerConfiguration.getAskForReplicatePeriod()).thenReturn(5000L);
 
         assertThat(workerService.isWorkerAllowedToAskReplicate(wallet)).isFalse();
@@ -269,13 +242,12 @@ class WorkerServiceTests {
     void shouldUpdateLastReplicateDemand() {
         String wallet = "wallet";
         Date lastDate = Date.from(Instant.now().minusSeconds(1));
-        Worker worker = Worker.builder()
-                .lastReplicateDemandDate(lastDate)
-                .build();
-        when(workerRepository.findByWalletAddress(wallet)).thenReturn(Optional.of(worker));
+        workerService.getWorkerStatsMap().computeIfAbsent(wallet, WorkerService.WorkerStats::new)
+                .setLastReplicateDemandDate(lastDate);
 
-        assertThat(workerService.updateLastReplicateDemandDate(wallet)
-                .get()
+        workerService.updateLastReplicateDemandDate(wallet);
+        assertThat(workerService.getWorkerStatsMap()
+                .get(wallet)
                 .getLastReplicateDemandDate())
                 .isAfter(lastDate);
     }
@@ -293,7 +265,7 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
+                .maxNbTasks(7)
                 .participatingChainTaskIds(new ArrayList<>(Arrays.asList("task1", "task2")))
                 .computingChainTaskIds(new ArrayList<>(Arrays.asList("task1", "task2")))
                 .build();
@@ -311,9 +283,32 @@ class WorkerServiceTests {
     }
 
     @Test
-    void shouldNotAddTaskIdToWorker() {
+    void shouldNotAddTaskIdToWorkerSinceUnknownWorker() {
         when(workerRepository.findByWalletAddress(Mockito.anyString())).thenReturn(Optional.empty());
         Optional<Worker> addedWorker = workerService.addChainTaskIdToWorker("task1", "0x1a69b2eb604db8eba185df03ea4f5288dcbbd248");
+        assertThat(addedWorker).isEmpty();
+    }
+
+    @Test
+    void shouldNotAddTaskIdToWorkerSinceCantAcceptMoreWorker() {
+        String workerName = "worker1";
+        String walletAddress = "0x1a69b2eb604db8eba185df03ea4f5288dcbbd248";
+        Worker existingWorker = Worker.builder()
+                .id("1")
+                .name(workerName)
+                .walletAddress(walletAddress)
+                .os("Linux")
+                .cpu("x86")
+                .cpuNb(3)
+                .maxNbTasks(2)
+                .participatingChainTaskIds(new ArrayList<>(Arrays.asList("task1", "task2")))
+                .computingChainTaskIds(new ArrayList<>(Arrays.asList("task1", "task2")))
+                .build();
+
+        when(workerRepository.findByWalletAddress(walletAddress)).thenReturn(Optional.of(existingWorker));
+        when(workerRepository.save(existingWorker)).thenReturn(existingWorker);
+
+        Optional<Worker> addedWorker = workerService.addChainTaskIdToWorker("task3", walletAddress);
         assertThat(addedWorker).isEmpty();
     }
 
@@ -373,7 +368,6 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .participatingChainTaskIds(new ArrayList<>(Arrays.asList("task1", "task2")))
                 .computingChainTaskIds(new ArrayList<>(Arrays.asList("task1", "task2")))
                 .build();
@@ -410,7 +404,6 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .participatingChainTaskIds(participatingIds)
                 .computingChainTaskIds(computingIds)
                 .build();
@@ -441,7 +434,6 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .participatingChainTaskIds(participatingIds)
                 .computingChainTaskIds(computingIds)
                 .build();
@@ -472,7 +464,6 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .participatingChainTaskIds(participatingIds)
                 .computingChainTaskIds(computingIds)
                 .build();
@@ -497,7 +488,6 @@ class WorkerServiceTests {
                 .os("Linux")
                 .cpu("x86")
                 .cpuNb(8)
-                .lastAliveDate(new Date())
                 .participatingChainTaskIds(participatingIds)
                 .computingChainTaskIds(computingIds)
                 .build();
@@ -518,18 +508,12 @@ class WorkerServiceTests {
     @Test
     void shouldGetLostWorkers() {
 
-        List<Worker> allWorkers = getDummyWorkers(3);
+        List<Worker> allWorkers = getDummyWorkers();
+
         List<Worker> lostWorkers = allWorkers.subList(1, 3);
-        when(workerRepository.findByLastAliveDateBefore(Mockito.any())).thenReturn(lostWorkers);
+        when(workerRepository.findByWalletAddressIn(Mockito.any())).thenReturn(lostWorkers);
 
         List<Worker> claimedLostWorkers = workerService.getLostWorkers();
-
-        // check findByLastAliveDateBefore was called with a date of one minute ago
-        ArgumentCaptor<Date> argument = ArgumentCaptor.forClass(Date.class);
-        Mockito.verify(workerRepository).findByLastAliveDateBefore(argument.capture());
-        long diff = (new Date()).getTime() - argument.getValue().getTime();
-        long diffInMinutes = TimeUnit.MILLISECONDS.toMinutes(diff);
-        assertThat(diffInMinutes).isEqualTo(1);
 
         // check the claimedLostWorkers are actually the lostWorkers
         assertThat(claimedLostWorkers)
@@ -539,27 +523,17 @@ class WorkerServiceTests {
 
     @Test
     void shouldNotFindLostWorkers() {
-
-        when(workerRepository.findByLastAliveDateBefore(Mockito.any())).thenReturn(Collections.emptyList());
-
         assertThat(workerService.getLostWorkers()).isEmpty();
     }
 
     @Test
     void shouldGetAliveWorkers() {
 
-        List<Worker> allWorkers = getDummyWorkers(3);
+        List<Worker> allWorkers = getDummyWorkers();
         List<Worker> aliveWorkers = allWorkers.subList(0, 1);
-        when(workerRepository.findByLastAliveDateAfter(Mockito.any())).thenReturn(aliveWorkers);
+        when(workerRepository.findByWalletAddressIn(Mockito.any())).thenReturn(aliveWorkers);
 
         List<Worker> claimedAliveWorkers = workerService.getAliveWorkers();
-
-        // check findByLastAliveDateAfter was called with a date of one minute ago
-        ArgumentCaptor<Date> argument = ArgumentCaptor.forClass(Date.class);
-        Mockito.verify(workerRepository).findByLastAliveDateAfter(argument.capture());
-        long diff = (new Date()).getTime() - argument.getValue().getTime();
-        long diffInMinutes = TimeUnit.MILLISECONDS.toMinutes(diff);
-        assertThat(diffInMinutes).isEqualTo(1);
 
         // check the claimedAliveWorkers are actually the aliveWorkers
         assertThat(claimedAliveWorkers)
@@ -569,9 +543,7 @@ class WorkerServiceTests {
 
     @Test
     void shouldNotFindAliveWorkers() {
-
-        when(workerRepository.findByLastAliveDateAfter(Mockito.any())).thenReturn(Collections.emptyList());
-
+        when(workerRepository.findByWalletAddressIn(Mockito.any())).thenReturn(Collections.emptyList());
         assertThat(workerService.getAliveWorkers()).isEmpty();
     }
 
@@ -583,19 +555,8 @@ class WorkerServiceTests {
                 3,
                 Arrays.asList("task1", "task2", "task3", "task4", "task5"),
                 Arrays.asList("task1", "task3"));
-        when(workerRepository.findByWalletAddress(walletAddress)).thenReturn(Optional.of(worker));
 
-        assertThat(workerService.canAcceptMoreWorks(walletAddress)).isTrue();
-    }
-
-    @Test
-    void shouldNotAcceptMoreWorksSinceWorkerNotFound() {
-        String walletAddress = "0x1a69b2eb604db8eba185df03ea4f5288dcbbd248";
-
-        when(workerRepository.findByWalletAddress(Mockito.any())).thenReturn(Optional.empty());
-
-        boolean canAccept = workerService.canAcceptMoreWorks(walletAddress);
-        assertThat(canAccept).isFalse();
+        assertThat(workerService.canAcceptMoreWorks(worker)).isTrue();
     }
 
     @Test
@@ -606,19 +567,19 @@ class WorkerServiceTests {
                 2,
                 Arrays.asList("task1", "task2", "task3", "task4"),
                 Arrays.asList("task1", "task3"));
-        when(workerRepository.findByWalletAddress(Mockito.anyString())).thenReturn(Optional.of(worker));
 
-        assertThat(workerService.canAcceptMoreWorks(walletAddress)).isFalse();
+        assertThat(workerService.canAcceptMoreWorks(worker)).isFalse();
     }
 
-    List<Worker> getDummyWorkers(int n) {
-
-        List<Worker> dummyWorkers = new ArrayList<>();
-
-        for (int i = 0; i < n; i++) {
-            dummyWorkers.add(Worker.builder().id(Integer.toString(i)).build());
-        }
-        return dummyWorkers;
+    List<Worker> getDummyWorkers() {
+        workerService.getWorkerStatsMap().computeIfAbsent("address1", WorkerService.WorkerStats::new).setLastAliveDate(new Date());
+        workerService.getWorkerStatsMap().computeIfAbsent("address2", WorkerService.WorkerStats::new).setLastAliveDate(addMinutesToDate(new Date(), -2));
+        workerService.getWorkerStatsMap().computeIfAbsent("address3", WorkerService.WorkerStats::new).setLastAliveDate(addMinutesToDate(new Date(), -3));
+        return List.of(
+                Worker.builder().id("1").walletAddress("address1").build(),
+                Worker.builder().id("2").walletAddress("address2").build(),
+                Worker.builder().id("3").walletAddress("address3").build()
+        );
     }
 
     Worker getDummyWorker(String walletAddress, int cpuNb, List<String> participatingIds, List<String> computingIds) {
@@ -645,7 +606,7 @@ class WorkerServiceTests {
                 4,
                 Arrays.asList("task1", "task2", "task3", "task4"),
                 List.of("task1"));//3 CPUs available
-        when(workerRepository.findByLastAliveDateAfter(any())).thenReturn(Arrays.asList(worker1, worker2));
+        when(workerRepository.findByWalletAddressIn(any())).thenReturn(Arrays.asList(worker1, worker2));
 
         assertThat(workerService.getAliveAvailableCpu()).isEqualTo(5);
     }
@@ -663,14 +624,14 @@ class WorkerServiceTests {
                 4,
                 Arrays.asList("task1", "task2", "task3", "task4"),
                 Arrays.asList("task1", "task2", "task3", "task4"));
-        when(workerRepository.findByLastAliveDateAfter(any())).thenReturn(Arrays.asList(worker1, worker2));
+        when(workerRepository.findByWalletAddressIn(any())).thenReturn(Arrays.asList(worker1, worker2));
 
         assertThat(workerService.getAliveAvailableCpu()).isZero();
     }
 
     @Test
     void shouldGetZeroAvailableCpuIfNoWorkerAlive() {
-        when(workerRepository.findByLastAliveDateAfter(any())).thenReturn(Collections.emptyList());
+        when(workerRepository.findByWalletAddressIn(any())).thenReturn(Collections.emptyList());
         assertThat(workerService.getAliveAvailableCpu()).isZero();
     }
 
@@ -687,7 +648,7 @@ class WorkerServiceTests {
                 .computingChainTaskIds(List.of("T4"))
                 .build();
         List<Worker> list = List.of(worker1, worker2);
-        when(workerRepository.findByLastAliveDateAfter(any())).thenReturn(list);
+        when(workerRepository.findByWalletAddressIn(any())).thenReturn(list);
         workerService.init();
         workerService.updateMetrics();
 
@@ -718,7 +679,7 @@ class WorkerServiceTests {
                 .gpuEnabled(false)
                 .build();
         List<Worker> list = List.of(worker1, worker2);
-        when(workerRepository.findByLastAliveDateAfter(any())).thenReturn(list);
+        when(workerRepository.findByWalletAddressIn(any())).thenReturn(list);
 
         assertThat(workerService.getAliveTotalGpu()).isEqualTo(1);
     }
@@ -736,7 +697,7 @@ class WorkerServiceTests {
                 .computingChainTaskIds(List.of("t1"))
                 .build();
         List<Worker> list = List.of(worker1, worker2);
-        when(workerRepository.findByLastAliveDateAfter(any())).thenReturn(list);
+        when(workerRepository.findByWalletAddressIn(any())).thenReturn(list);
 
         assertThat(workerService.getAliveAvailableGpu()).isEqualTo(1);
     }